sdk: split query_verbs into module directory (5/9: extract state)
Move state_status, fingerprint, and helpers to query_verbs/state.rs.
Visibility: FingerprintOptions::scope fn -> pub(crate) (used from tests module).
